+++ packages/flame/CHANGELOG.md
@@ -51,6 +51,9 @@
  - `TextBoxComponent` can have customizable `pixelRatio`
  - Add `ContainsAtLeastMockCanvas` to facilitate testing with `MockCanvas`
  - Support for `drawImage` for `MockCanvas`
+ - `Game` is now a `Component`
+ - `ComponentEffect` is now a `Component`
+ - `HasGameRef` can now operate independently from `Game`
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps